#이정욱 리뷰게시판 글작성 스마트에디터 DB연동 기능구현

+++ branches/teamB_project_DEV-1.0/src/main/java/kr/ync/project/domain/Review_regiVO.java
@@ -16,7 +16,7 @@ public class Review_regiVO {
 	private String USER_ID;	//	작성자 [fk]
 	private int PRO_NUM; //		제품 번호 [fk]
 	private Date REVIEW_DATE; //  작성일
-	private int REVIEW_GRADE;  //  평점
+	private float REVIEW_GRADE;  //  평점
 	private String REVIEW_CONT; // 	내용 
 	private String REVIEW_IMAGE; // 이미지 번호
 	private int REVIEW_HITS; // 조회수
